+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 
 ### Changed
 
+- 🛂(backend) let users allowed to comment list each other's accesses
 - ⬆️(backend) upgrade celery to version 5.6.3
 - ⚡️(backend) stop using LEFT(value, LENGTH(path)) in sql queries
 - 🚚(project) switch docspec image to ghcr.io/docspec/api #2553
@@ -166,7 +167,6 @@ and this project adheres to
 
 ### Changed
 
-- 🛂(backend) make document access list visible to all collaborators
 - 👷(CI) remove test-e2e-other-browser job #2404
 - ♿️(frontend) use heading element for pinned documents section title #2380
 - ♿️(frontend) use anchor links for table of contents entries #2390